Refactor CustomerForm
